pyproject.toml
[tool.poetry]
name = "pz"
version = "0.1.0"
description = ""
authors = ["Vitaliy Drozdov <vitdrzdv@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12, <3.13"
Django = "^4.2.5"
python-dotenv = "~1.0"
psycopg2-binary = "~2.9"
djangorestframework = "3.15.1"
faker = "^28.4.1"
redis = "^5.0.8"
hiredis = "^3.0.0"
gunicorn = "^23.0.0"
drf-yasg = "^1.21.7"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
black = "^24.4.0"
django-debug-toolbar = "^4.2.0"



[tool.isort]
profile = "black"
skip_glob = ["**/migrations/*.py"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
